+++ internal/runner/vars.go
@@ -3,4 +3,5 @@ package runner
 var (
 	google204 = "http://clients3.google.com/generate_204"
 	doUpdate  bool
+	version   bool
 )