+++ internal/updater/changes.go
@@ -1,0 +1,6 @@
+package updater
+
+type changes struct {
+	Name string `json:"name"`
+	Body string `json:"body"`
+}