+++ internal/updater/vars.go
@@ -1,0 +1,6 @@
+package updater
+
+var (
+	sure      bool
+	binaryURL = "https://github.com/kitabisa/APP/releases/download/VVER/APP_VER_OS_ARCH"
+)